showcmd instead of cursor shape in op-pending mode
This solution was hacky to begin with, and it had some unintended visual
side effects such as the appearance of briefly entering insert mode upon
leaving insert mode. In addition, the cursor shape does not change to an
underline unless Vim is kept idle for a bit due to this code running on
a timer. Setting showcmd achieves the goal of giving a visual indication
of being in operator pending mode.