Overhaul Termite config file management
vga.config has been moved to a file vga.config.in meant to supplement
the default configuration file, and this has been paired with wrappers
for Termite and Neofetch (the latter because Neofetch reads the default
config file for information about the terminal). Additionally,
vga.config.in has been updated for version 2.0 of the Ultimate Oldschool
PC Font Pack.